Modified merges and added ajax and block.
There are now only 2 merge functions, mergeBlocking and
mergeNonBlocking. See the comments and the unit tests for
usage details.

Added ajax, which takes in jQuery ajax arguments and calls
jQuery.ajax with them and immediately returns a new
EventStream. It also takes in an additional function to
return a new EventStream with the jQuery ajax results.

Added block, which takes in a function that takes in the
next value and a callback. The function is free to take as
long as it wants to complete, and then call the callback
with the next value to allow the EventStream to continue.